[NFC] Replace and Refactor TwoLevelMap
Add loads of documentation comments and clean up a lot of the interfaces to things.
